+++ packages/config/src/ProfileInfo.ts
@@ -19,7 +19,7 @@ import * as lodash from "lodash";
 // for ProfileInfo structures
 import { IProfArgAttrs } from "./doc/IProfArgAttrs";
 import { IProfAttrs } from "./doc/IProfAttrs";
-import { IProfLoc, ProfLocType } from "./doc/IProfLoc";
+import { IProfLoc, IProfLocOsLoc, IProfLocOsLocLayer, ProfLocType } from "./doc/IProfLoc";
 import { IProfMergeArgOpts } from "./doc/IProfMergeArgOpts";
 import { IProfMergedArg } from "./doc/IProfMergedArg";
 import { IProfOpts } from "./doc/IProfOpts";
@@ -197,7 +197,7 @@ export class ProfileInfo {
         }
 
         const mergedArgs = this.mergeArgsForProfile(desiredProfile, { getSecureVals: false });
-        if (!(await this.updateKnownProperty({ mergedArgs, property: options.property, value: options.value, setSecure: options.setSecure }))) {
+        if (!(await this.updateKnownProperty({ ...options, mergedArgs, osLocInfo: this.getOsLocInfo(desiredProfile)[0] }))) {
             if (this.usingTeamConfig) {
                 // Check to see if loadedConfig already contains the schema for the specified profile type
                 if (ImperativeConfig.instance.loadedConfig?.profiles?.find(p => p.type === options.profileType)?.schema == null ||
@@ -274,8 +274,19 @@ export class ProfileInfo {
                 break;
             }
             case ProfLocType.TEAM_CONFIG: {
+                let oldLayer: IProfLocOsLocLayer;
+                if (options.osLocInfo) {
+                    const layer = this.getTeamConfig().layerActive();
+                    oldLayer = { user: layer.user, global: layer.global };
+                    this.getTeamConfig().api.layers.activate(options.osLocInfo.user, options.osLocInfo.global);
+                }
+
                 this.getTeamConfig().set(toUpdate.argLoc.jsonLoc, options.value, { secure: options.setSecure });
                 await this.getTeamConfig().save(false);
+
+                if (oldLayer) {
+                    this.getTeamConfig().api.layers.activate(oldLayer.user, oldLayer.global);
+                }
                 break;
             }
             case ProfLocType.ENV:
@@ -893,6 +904,28 @@ export class ProfileInfo {
     public get usingTeamConfig(): boolean {
         this.ensureReadFromDisk();
         return this.mUsingTeamConfig;
+    }
+
+    /**
+     * Gather information about the paths in osLoc
+     * @param profile Profile attributes gathered from getAllProfiles
+     */
+    public getOsLocInfo(profile: IProfAttrs): IProfLocOsLoc[] {
+        const osLoc = profile?.profLoc.osLoc;
+        if (!osLoc?.length) return undefined;
+        if (profile.profLoc.locType === ProfLocType.TEAM_CONFIG) {
+            const ret: IProfLocOsLoc[] = [];
+            for (const loc of osLoc) {
+                for (const layer of this.mLoadedConfig.layers) {
+                    if (layer.path === loc) {
+                        // we found the config layer matching osLoc
+                        ret.push({ name: profile.profName, path: loc, user: layer.user, global: layer.global });
+                    }
+                }
+            }
+            return ret;
+        }
+        return [{ name: profile.profName, path: profile.profLoc.osLoc[0], user: undefined, global: undefined }];
     }
 
     /**